+++ packages/at_client/lib/at_client_mixins.dart
@@ -2,4 +2,5 @@ export 'package:at_client/src/mixins/at_client_bindings.dart';
 export 'package:at_client/src/mixins/apkam_signing.dart';
 export 'package:at_client/src/mixins/at_client_envelope_signer.dart';
 export 'package:at_client/src/mixins/envelope_signing.dart';
+export 'package:at_client/src/enroll/self_retrofit.dart';
 export 'package:at_client/src/secret_sharing/secret_sharing.dart';